Refactor plots
The voltage domain plots were removed since they aren't relevant to the
fit anymore. The acceleration plots were removed since they are almost
always noise and aren't helpful for troubleshooting.

The GUI layout was reorganized so all the plots fit in the main display
and the only windows with a scrollbar are the generator and program log.

Instances of "units" are now replaced with an abbreviated unit name so
they use less horizontal space.

Fixes #269.